Consolidate name of feature file
