+++ packages/technical-features/application/electron-main/index.ts
@@ -3,6 +3,6 @@ import { overrideSideEffectsWithFakes } from "./src/override-side-effects-with-f
 export { beforeElectronIsReadyInjectionToken } from "./src/start-application/timeslots/before-electron-is-ready-injection-token";
 export { beforeAnythingInjectionToken } from "./src/start-application/timeslots/before-anything-injection-token";
 
-export { applicationFeatureForElectronMain } from "./src/applicationFeatureForElectronMain";
+export { applicationFeatureForElectronMain } from "./src/feature";
 
 export const testUtils = { overrideSideEffectsWithFakes }
