+++ packages/technical-features/application/electron-main/src/start-application/when-app-is-ready/when-app-is-ready.test.ts
@@ -1,5 +1,5 @@
 import { createContainer, DiContainer } from "@ogre-tools/injectable";
-import { applicationFeatureForElectronMain } from "../../applicationFeatureForElectronMain";
+import { applicationFeatureForElectronMain } from "../../feature";
 import { registerFeature } from "@k8slens/feature-core";
 import whenAppIsReadyInjectable from "../when-app-is-ready.injectable";
 import { getPromiseStatus } from "@ogre-tools/test-utils";
